--- ST/Evgen_Gal/Project/src/components/catalog.js
@@ -1,52 +1,0 @@
-import CatalogItem from "./catalogItem.js";
-export default class Catalog {
-  constructor(container, url, basket) {
-    this.items = [];
-    this.container = document.querySelector(container);
-    this.url = url;
-    this.basket = basket;
-    this._init();
-  }
-
-  _init() {
-    this._get(this.url)
-      .then((arr) => {
-        this.items = arr;
-      })
-      .finally(() => {
-        this._render();
-        this._handleActions();
-      });
-  }
-  _get(url) {
-    return fetch(url).then((d) => d.json());
-  }
-  _fillCatalog() {
-    //Инкапсуляция (условная для JS)
-    this.items = getArrayOfObjects();
-  }
-  _render() {
-    let htmlStr = "";
-    this.items.forEach((item) => {
-      htmlStr += new CatalogItem(item).render();
-    });
-    this.container.innerHTML = htmlStr;
-  }
-  _handleActions() {
-    this.container.addEventListener("click", (ev) => {
-      if (ev.target.name == "add") {
-        let dataset = ev.target.dataset;
-        this.basket.add(this._createNewItem(dataset));
-      }
-    });
-  }
-  _createNewItem(dataset) {
-    return {
-      productId: dataset.id,
-      productName: dataset.name,
-      productImg: dataset.img,
-      productPrice: +dataset.price,
-      amount: 1,
-    };
-  }
-}
