+++ ST/Evgen_Gal/Project/src/components/index.js
@@ -1,12 +1,6 @@
-import Basket from "./basket";
-import Catalog from "./catalog";
+import { Basket, Catalog } from "./LISTS";
 
 export default () => {
-  let catalogUrl =
-    "https://raw.githubusercontent.com/kellolo/static/master/JSON/catalog.json";
-  let basketUrl =
-    "https://raw.githubusercontent.com/kellolo/static/master/JSON/basket.json";
-
-  let basket = new Basket("#basket", basketUrl);
-  let catalog = new Catalog("#catalog", catalogUrl, basket);
+  let basket = new Basket();
+  let catalog = new Catalog(basket);
 };
