+++ ST/Evgen_Gal/Project/src/index.js
@@ -2,6 +2,6 @@ import bootstrap from "bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./layout/styles/styles.css";
 
-import app from "./components"; // импорт приложение из "./components"; при одноименном назавнии подгружает его автоматом
+import app from "./components/index"; // импорт приложение из "./components"; при одноименном назавнии подгружает его автоматом
 
 app();
